+++ demo/src/test/java/com/javaobjectcopy/demo/shallowcopy/ShoppingCartShallowCopyUnitTest.java
@@ -1,5 +1,6 @@
 package com.javaobjectcopy.demo.shallowcopy;
 
+import com.javaobjectcopy.demo.deepcopy.Item1;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -9,7 +10,7 @@ class ShoppingCartShallowCopyUnitTest {
     void whenShallowClonedObjectIsUpdated_thenChangeShouldReflectInOriginal() throws CloneNotSupportedException {
         ShoppingCartShallowCopy original = new ShoppingCartShallowCopy(2, new Item1("soap"));
 
-        ShoppingCartShallowCopy cloned = (ShoppingCartShallowCopy) original.tryClone();
+        ShoppingCartShallowCopy cloned = (ShoppingCartShallowCopy) original.clone();
 
         Assertions.assertNotNull(cloned);
         Assertions.assertEquals(original.getItem().getName(), cloned.getItem().getName(), "After Cloning:\nItem name should be equal");
